Day 11 / Part 2 https://adventofcode.com/2022/day/11#part2

+++ src/main/kotlin/days/day11/Item.kt
@@ -1,3 +1,3 @@
 package days.day11
 
-data class Item(var worry: Int)
+data class Item(var worry: Long)
