+++ src/main/kotlin/days/day11/Operation.kt
@@ -1,8 +1,8 @@
 package days.day11
 
 class Operation(
-    private val operand1: Int?,
-    private val operand2: Int?,
+    private val operand1: Long?,
+    private val operand2: Long?,
     private val operator: Char,
 ) {
     companion object {
@@ -12,7 +12,7 @@ class Operation(
         const val DIVIDE = '/'
     }
 
-    fun calculate(old: Int): Int {
+    fun calculate(old: Long): Long {
         val a = (operand1 ?: old)
         val b = (operand2 ?: old)
